Flytt Hendelsesvisning til egen component
